==> 3_GetCSV_Charges.ipynb <==
import matplotlib as mpl
from mpl_toolkits.mplot3d import Axes3D
import numpy as np
import matplotlib.pyplot as plt
import csv
import pandas as pd
from sklearn.preprocessing import StandardScaler

# ----------------

def crystal(i):
    switcher={
        45: "Li", 
        44: "Cu",
        43: "Mn",
        42: "Al",            
        41: "Au",            
        40: "Ca",      
        39: "Cr",        
        38: "Fe",            
        37: "Ge",            
        36: "Ir",      
        35: "Pd",        
        34: "Pt",            
        33: "Ru",            
        32: "Si",      
        31: "Tc",
        30: "Ti",
        29: "V",
        28: "W",
        27: "Mg",
        26: "Co",
        25: "Sc", 
        24: "Hg",
        23: "Mo", 
        22: "Os",
        21: "Rh",
        20: "Ag",
        19: "Cd",
        18: "Hf",
        17: "K",
        16: "Na",
        15: "Nb",
        14: "Ni",
        13: "Re",
        12: "Sb",
        11: "Sr",
        10: "Ta",
        9: "Y",
        8: "Zr",
        7: "Ba",
        6: "Cs",
        5: "In",
        4: "Rb",
        3: "Sn",
        2: "Pb",
        1: "Tl",
    }
    return switcher.get(i,"Invalid crystal")


def non_pure_crystal(i):
    switcher={
		11: "CrH2",
        10: "VH2",
		9: "TiH2",
        8: "ScH3",
        7: "NbH2",
        6: "MgH2",
        5: "CaH3",
        4: "Na2LiAlH6",
        3: "K2NaAlH6",
        2: "K2LiAlH6",
        1: "AlH3",
    }
    return switcher.get(i,"Invalid crystal")

#df = pd.DataFrame(list())
df1 = pd.DataFrame(list())

for pp in range (1,46):  #18 pure crystal #12 non pure crystal
    
    
    crystalname=crystal(pp)
    
    #source = "/home/massad/Documents/Linux_PHD/SPIN_CHECK_CORRECT_2/"+crystalname+"_copy/Density/"+crystalname+".H_charge_diff_OCTA.dat"
    #source1 = "/home/massad/Documents/Linux_PHD/SPIN_CHECK_CORRECT_2/"+crystalname+"_copy/Density/"+crystalname+".without_H_charge_diff_OCTA.dat"
    
    #source = "/home/massad/Documents/Linux_PHD/SPIN_CHECK_CORRECT_2/"+crystalname+"_tetra_copy/Density/"+crystalname+".H_charge_diff_TETRA.dat"
    source1 = "/home/massad/Documents/Linux_PHD/SPIN_CHECK_CORRECT_2/"+crystalname+"_tetra_copy/Density/"+crystalname+".without_H_charge_diff_TETRA.dat"
    
    #print("Current source,",source)
    print("Current source,",source1)
    
    #text_file = open(source, "r")
    text_file1 = open(source1, "r")
    
    #lines = text_file.read()
    lines1 = text_file1.read()
    
    #a=lines.split("\n")
    a=lines1.split("\n")

    b=a[0]
    c=b.split(" ")
    del c[0]
    d=[]
    for i in range(0,len(a)):
        b=a[i]
        c=b.split(" ")
        #print(c)
        del c[0]
        for j in range(0,len(c)):
            d.append(c[j])

    print('Data length,',len(d))
    print('')
    print('Type after acquiring,',type(d))
    e=np.array(d)
    e=e.astype(np.float64)
    #print(e[64],e[111],e[1134])
    print('')
    print('This is the final vector,',e)
    print('')
    #print('Type,',type(e))
    print('')
    print('Shape,',e.shape)
    print('')

    print("A test has been made, this data is in the 'raveled' shape of n_x x n_y x n_z")


#----- writing on CSV

    print(len(e))
    print(e[0])

    #df.to_csv('Charge_diff_BulkH_OCTA.csv') # it creates it
    
    #df[crystalname] = pd.Series(e)

    #df.to_csv("Charge_diff_BulkH_OCTA.csv", index=False)
    
    
    #df1.to_csv('Charge_diff_Bulk_OCTA.csv') # it creates it
    
    #df1[crystalname] = pd.Series(e)

    #df1.to_csv("Charge_diff_Bulk_OCTA.csv", index=False)
    
    
    #df.to_csv('Charge_diff_BulkH_TETRA.csv') # it creates it
    
    #df[crystalname] = pd.Series(e)

    #df.to_csv("Charge_diff_BulkH_TETRA.csv", index=False)
    
    
    #df1.to_csv('Charge_diff_Bulk_TETRA.csv') # it creates it
    
    #df1[crystalname] = pd.Series(e)

    #df1.to_csv("Charge_diff_Bulk_TETRA.csv", index=False)
    
    # --------------------
    

import matplotlib as mpl
from mpl_toolkits.mplot3d import Axes3D
import numpy as np
import matplotlib.pyplot as plt
import csv
import pandas as pd
from sklearn.preprocessing import StandardScaler

# ----------------

def crystal(i):
    switcher={
        45: "Li", 
        44: "Cu",
        43: "Mn",
        42: "Al",            
        41: "Au",            
        40: "Ca",      
        39: "Cr",        
        38: "Fe",#no          
        37: "Ge",            
        36: "Ir",      
        35: "Pd",        
        34: "Pt",            
        33: "Ru",            
        32: "Si", #no     
        31: "Tc",
        30: "Ti",
        29: "V",
        28: "W",#no
        27: "Mg",
        26: "Co",#no
        25: "Sc", 
        24: "Hg",#no
        23: "Mo", 
        22: "Os",
        21: "Rh",
        20: "Ag",
        19: "Cd",
        18: "Hf",#no
        17: "K",
        16: "Na",
        15: "Nb",
        14: "Ni",#no
        13: "Re",
        12: "Sb",#no
        11: "Sr",
        10: "Ta",
        9: "Y",
        8: "Zr",
        7: "Ba", #no 
        6: "Cs", #no
        5: "In",
        4: "Rb",
        3: "Sn",
        2: "Pb",
        1: "Tl",
    }
    return switcher.get(i,"Invalid crystal")

df = pd.DataFrame(list())

for pp in range (1,46): #OCTA+H
    
    if pp==7 or pp==26 or pp==38 or pp==14 or pp==28 or pp==18 or pp==24:
        continue
        
    if pp==6 or pp==32 or pp==12:
        continue
    
    crystalname=crystal(pp)
    source = "./"+crystalname+"_copy/Density/"+crystalname+".H.supercell_charge_diff_OCTA.dat"
    print("Current source,",source)
    
    text_file = open(source, "r")
    lines = text_file.read()
    a=lines.split("\n")
    
    b=a[0]
    c=b.split(" ")
    del c[0]
    d=[]
    for i in range(0,len(a)):
        b=a[i]
        c=b.split(" ")
        #print(c)
        del c[0]
        for j in range(0,len(c)):
            d.append(c[j])

    print('Data length,',len(d))
    print('')
    print('Type after acquiring,',type(d))
    e=np.array(d)
    e=e.astype(np.float64)
    #print(e[64],e[111],e[1134])
    print('')
    print('This is the final vector,',e)
    print('')
    #print('Type,',type(e))
    print('')
    print('Shape,',e.shape)
    print('')

    print("A test has been made, this data is in the 'raveled' shape of n_x x n_y x n_z")

#----- writing on CSV

    print(len(e))
    print(e[0])
    
    df[crystalname] = pd.Series(e)

    df.to_csv("Charge_diff_BulkH_OCTA.csv", index=False)
    
    
    #--------------------------------------------------------------

for pp in range (1,46): #OCTA
    
    if pp==7 or pp==26 or pp==38 or pp==14 or pp==28 or pp==18 or pp==24:
        continue
        
    if pp==6 or pp==32 or pp==12:
        continue
        
    crystalname=crystal(pp)
    source = "./"+crystalname+"_copy/Density/"+crystalname+".supercell_charge_diff_OCTA.dat"
    print("Current source,",source)
    
    text_file = open(source, "r")
    lines = text_file.read()
    a=lines.split("\n")
    
    b=a[0]
    c=b.split(" ")
    del c[0]
    d=[]
    for i in range(0,len(a)):
        b=a[i]
        c=b.split(" ")
        #print(c)
        del c[0]
        for j in range(0,len(c)):
            d.append(c[j])

    print('Data length,',len(d))
    print('')
    print('Type after acquiring,',type(d))
    e=np.array(d)
    e=e.astype(np.float64)
    #print(e[64],e[111],e[1134])
    print('')
    print('This is the final vector,',e)
    print('')
    #print('Type,',type(e))
    print('')
    print('Shape,',e.shape)
    print('')

    print("A test has been made, this data is in the 'raveled' shape of n_x x n_y x n_z")

#----- writing on CSV

    print(len(e))
    print(e[0])
    
    df[crystalname] = pd.Series(e)

    df.to_csv("Charge_diff_Bulk_OCTA.csv", index=False)    
    
    #-----------------------------------------------------
    
for pp in range (1,46): #TETRA+H
    
    if pp==7 or pp==26 or pp==38 or pp==14 or pp==28 or pp==18 or pp==24:
        continue
        
    if pp==6 or pp==32 or pp==12:
        continue
        
    crystalname=crystal(pp)
    source = "./"+crystalname+"_tetra_copy/Density/"+crystalname+".H.supercell_charge_diff_TETRA.dat"
    print("Current source,",source)
    
    text_file = open(source, "r")
    lines = text_file.read()
    a=lines.split("\n")
    
    b=a[0]
    c=b.split(" ")
    del c[0]
    d=[]
    for i in range(0,len(a)):
        b=a[i]
        c=b.split(" ")
        #print(c)
        del c[0]
        for j in range(0,len(c)):
            d.append(c[j])

    print('Data length,',len(d))
    print('')
    print('Type after acquiring,',type(d))
    e=np.array(d)
    e=e.astype(np.float64)
    #print(e[64],e[111],e[1134])
    print('')
    print('This is the final vector,',e)
    print('')
    #print('Type,',type(e))
    print('')
    print('Shape,',e.shape)
    print('')

    print("A test has been made, this data is in the 'raveled' shape of n_x x n_y x n_z")

#----- writing on CSV

    print(len(e))
    print(e[0])
    
    df[crystalname] = pd.Series(e)

    df.to_csv("Charge_diff_BulkH_TETRA.csv", index=False)    
    
    #-----------------------------------------------------

for pp in range (1,46): #TETRA
    
    if pp==7 or pp==26 or pp==38 or pp==14 or pp==28 or pp==18 or pp==24:
        continue
        
    if pp==6 or pp==32 or pp==12:
        continue
        
    crystalname=crystal(pp)
    source = "./"+crystalname+"_tetra_copy/Density/"+crystalname+".supercell_charge_diff_TETRA.dat"
    print("Current source,",source)
    
    text_file = open(source, "r")
    lines = text_file.read()
    a=lines.split("\n")
    
    b=a[0]
    c=b.split(" ")
    del c[0]
    d=[]
    for i in range(0,len(a)):
        b=a[i]
        c=b.split(" ")
        #print(c)
        del c[0]
        for j in range(0,len(c)):
            d.append(c[j])

    print('Data length,',len(d))
    print('')
    print('Type after acquiring,',type(d))
    e=np.array(d)
    e=e.astype(np.float64)
    #print(e[64],e[111],e[1134])
    print('')
    print('This is the final vector,',e)
    print('')
    #print('Type,',type(e))
    print('')
    print('Shape,',e.shape)
    print('')

    print("A test has been made, this data is in the 'raveled' shape of n_x x n_y x n_z")

#----- writing on CSV

    print(len(e))
    print(e[0])
    
    df[crystalname] = pd.Series(e)

    df.to_csv("Charge_diff_Bulk_TETRA.csv", index=False)    
    
    #-----------------------------------------------------
