==> 4_Remeshing_Charges_TETRA.ipynb <==
import matplotlib as mpl
from mpl_toolkits.mplot3d import Axes3D
import numpy as np
# import matplotlib.pyplot as plt
import csv
import pandas as pd
from sklearn.preprocessing import StandardScaler
from pyrho.pgrid import PGrid
from pyrho.vis.scatter import get_scatter_plot
from scipy.interpolate import RegularGridInterpolator

def crystal(i):
    switcher={
        45: "Li", 
        44: "Cu",
        43: "Mn",
        42: "Al",            
        41: "Au",            
        40: "Ca",      
        39: "Cr",        
        38: "Fe",            
        37: "Ge",            
        36: "Ir",      
        35: "Pd",        
        34: "Pt",            
        33: "Ru",            
        32: "Si",      
        31: "Tc",
        30: "Ti",
        29: "V",
        28: "W",
        27: "Mg",
        26: "Co",
        25: "Sc", 
        24: "Hg",
        23: "Mo", 
        22: "Os",
        21: "Rh",
        20: "Ag",
        19: "Cd",
        18: "Hf",
        17: "K",
        16: "Na",
        15: "Nb",
        14: "Ni",
        13: "Re",
        12: "Sb",
        11: "Sr",
        10: "Ta",
        9: "Y",
        8: "Zr",
        7: "Ba",
        6: "Cs",
        5: "In",
        4: "Rb",
        3: "Sn",
        2: "Pb",
        1: "Tl",
    }
    return switcher.get(i,"Invalid crystal")

def non_pure_crystal(i):
    switcher={
		11: "CrH2",
        10: "VH2",
		9: "TiH2",
        8: "ScH3",
        7: "NbH2",
        6: "MgH2",
        5: "CaH3",
        4: "Na2LiAlH6",
        3: "K2NaAlH6",
        2: "K2LiAlH6",
        1: "AlH3",
    }
    return switcher.get(i,"Invalid crystal")

df = pd.read_csv("Charge_diff_BulkH_TETRA.csv")
df1 = pd.read_csv("Charge_diff_Bulk_TETRA.csv")

#df = pd.read_csv("Charge_diff_BulkH_TETRA.csv")
#df1 = pd.read_csv("Charge_diff_Bulk_TETRA.csv")

#df = pd.read_csv("Charge_diff_NonPure_BulkH_.csv")
#df1 = pd.read_csv("Charge_diff_NonPure_Bulk_.csv")

df2 = pd.DataFrame(list())
#df3 = pd.DataFrame(list())



# ------- REMESHING


for pp in range (1,46):  #18 pure crystal #12 non pure crystal

    # if pp==7 or pp==26 or pp==38 or pp==14 or pp==28 or pp==18 or pp==24:
    #     continue
        
    # if pp==6 or pp==32 or pp==12:
    #     continue
        
    # crystalname=crystal(pp)

    # print("")
    # print("Crystal name:",crystalname)
    # print("")

    # column = df[crystalname]
    # column = np.array(column)
    # column = column[np.logical_not(np.isnan(column))]
    
    # column1 = df1[crystalname]
    # column1 = np.array(column1)
    # column1 = column1[np.logical_not(np.isnan(column1))]
    
    # dimension=round(len(column)**(1/3))

    # dimension1=round(len(column1)**(1/3))
    
    # print("")
    # print("Dimension of column and column1:",dimension, dimension1)
    # print("")
    
    
    # column_rshp=column.reshape((dimension,dimension,dimension))
    # column1_rshp=column1.reshape((dimension1,dimension1,dimension1))
    
    # x = np.arange(column_rshp.shape[0])#[:, None, None]
    # y = np.arange(column_rshp.shape[1])#[None, :, None]
    # z = np.arange(column_rshp.shape[2])#[None, None, :]
    # x, y, z = np.broadcast_arrays(x, y, z)
    
    # x1 = np.arange(column1_rshp.shape[0])#[:, None, None]
    # y1 = np.arange(column1_rshp.shape[1])#[None, :, None]
    # z1 = np.arange(column1_rshp.shape[2])#[None, None, :]
    # x1, y1, z1 = np.broadcast_arrays(x1, y1, z1)
    
    # X0=np.arange(0,101,1,dtype=int)
    # Y0=np.arange(0,101,1,dtype=int)
    # Z0=np.arange(0,101,1,dtype=int)
    
    # column_remesh=ReGrid(column_rshp,x,y,z,X0,Y0,Z0)
    # column1_remesh=ReGrid(column1_rshp,x1,y1,z1,X0,Y0,Z0)
    
    # column_final=column_remesh.ravel()
    # column1_final=column1_remesh.ravel()

    # df2[crystalname+"_Difference"] = pd.Series(column_final-column1_final)
    # df2.to_csv("Final_Differences_Only_OCTA.csv", index=False)


    #----------------------------------

    if pp==7 or pp==26 or pp==38 or pp==14 or pp==28 or pp==18 or pp==24:
        continue
        
    if pp==6 or pp==32 or pp==12:
        continue
        
    crystalname=crystal(pp)
    
    print("")
    print("Crystal name:",crystalname)
    print("")

    column = df[crystalname]
    column = np.array(column)
    column = column[np.logical_not(np.isnan(column))]
    
    column1 = df1[crystalname]
    column1 = np.array(column1)
    column1 = column1[np.logical_not(np.isnan(column1))]

    dimension=round(len(column)**(1/3))

    dimension1=round(len(column1)**(1/3))
    
    print("")
    print("Dimension of column and column1:",dimension, dimension1)
    print("")
    
    column_rshp=column.reshape((dimension,dimension,dimension))
    column1_rshp=column1.reshape((dimension1,dimension1,dimension1))

    base='/home/massad/Documents/Linux_PHD/SUPERCELLS/'+crystalname+'_tetra_copy/Density/'

    myfile = open(base+crystalname+'.H.supercell_charge_diff.xsf', "r")

    import re
    cc=0
    lattice=[]
    while myfile:
        line  = myfile.readline()
        cc+=1
        # print(cc)
        if cc>2:
            # lattice.append(line)
            lattice.append([float(s) for s in re.findall(r"[-+]?[.]?[\d]+(?:,\d\d\d)*[\.]?\d*(?:[eE][-+]?\d+)?", line)])
        if cc>4:
            break
    myfile.close()

    print(lattice)

    # print(lattice[2][2])

    pg2d_diff = PGrid(grid_data=column_rshp, lattice=[[lattice[0][0], lattice[0][1], lattice[0][2]], [lattice[1][0], lattice[1][1], lattice[1][2]], [lattice[2][0], lattice[2][1], lattice[2][2]]])
    pg_2x_diff = pg2d_diff.get_transformed(sc_mat=[[1,0,0],[0,1,0],[0,0,1]],grid_out=[101, 101, 101], origin=[0.0,0.0,0.0])
    diffcharge_rgrd=pg_2x_diff.grid_data
    diffcharge_rgrd_flt=diffcharge_rgrd.ravel()

    pg2d_diff_1 = PGrid(grid_data=column1_rshp, lattice=[[lattice[0][0], lattice[0][1], lattice[0][2]], [lattice[1][0], lattice[1][1], lattice[1][2]], [lattice[2][0], lattice[2][1], lattice[2][2]]])
    pg_2x_diff_1 = pg2d_diff_1.get_transformed(sc_mat=[[1,0,0],[0,1,0],[0,0,1]],grid_out=[101, 101, 101], origin=[0.0,0.0,0.0])
    diffcharge_rgrd_1=pg_2x_diff_1.grid_data
    diffcharge_rgrd_flt_1=diffcharge_rgrd_1.ravel()

    df2[crystalname+"_Difference"] = pd.Series(diffcharge_rgrd_flt-diffcharge_rgrd_flt_1)
    df2.to_csv("Final_Differences_Only_TETRA.csv", index=False)
