==> 11_Side_Dataset_OCTA.ipynb <==
from mendeleev import element
#import pymatgen.core as mg
import matplotlib.pyplot as plt
#import matplotlib as mpl
#from mpl_toolkits.mplot3d import Axes3D
import numpy as np
#import matplotlib.pyplot as plt
import csv
import pandas as pd
from sklearn.preprocessing import StandardScaler
from sklearn.decomposition import PCA
#import quadpy
from tqdm.notebook import tqdm, trange
import time    # to be used in loop iterations
from joblib import Parallel, delayed, parallel_backend
from scipy import integrate
import time
from datetime import timedelta
import plotly as plotly
import plotly.express as px
#from plotly.offline import init_notebook_mode
#init_notebook_mode(connected=True)  
import plotly.io as pio
pio.renderers.default = 'iframe'
import plotly.graph_objects as go
import scipy.signal
from scipy.signal import argrelextrema
from scipy.signal import find_peaks

import os

def crystal(i):
    switcher={
        45: "Li", 
        44: "Cu",
        43: "Mn",
        42: "Al",            
        41: "Au",            
        40: "Ca",      
        39: "Cr",        
        38: "Fe",            
        37: "Ge",            
        36: "Ir",      
        35: "Pd",        
        34: "Pt",            
        33: "Ru",            
        32: "Si",      
        31: "Tc",
        30: "Ti",
        29: "V",
        28: "W",
        27: "Mg",
        26: "Co",
        25: "Sc", 
        24: "Hg",
        23: "Mo", 
        22: "Os",
        21: "Rh",
        20: "Ag",
        19: "Cd",
        18: "Hf",
        17: "K",
        16: "Na",
        15: "Nb",
        14: "Ni",
        13: "Re",
        12: "Sb",
        11: "Sr",
        10: "Ta",
        9: "Y",
        8: "Zr",
        7: "Ba",
        6: "Cs",
        5: "In",
        4: "Rb",
        3: "Sn",
        2: "Pb",
        1: "Tl",
    }
    return switcher.get(i,"Invalid crystal")

import plotly.express as px
from plotly.offline import init_notebook_mode, iplot
init_notebook_mode(connected=True)  
import plotly.io as pio
pio.renderers.default = 'iframe'

# element('Si').ionic_radii[0].ionic_radius

suffix = "OCTA"
df_intch=pd.read_csv("./NotIntCharge_"+suffix+".csv") #ok
df_bader=pd.read_csv("./ALL_BADER_"+suffix+".csv") #ok
df_lattice=pd.read_csv("./ALL_LATTICE.csv") #ok
df_structural=pd.read_csv("./ALL_STRUCTURAL.csv") #ok
df_valence=pd.read_csv("./ALL_VALENCE_"+suffix+".csv") #ok

data=[]
crystalnames_df=[]
    
counter=0
for pp in range (1,46):
    
    if pp==7 or pp==26 or pp==38 or pp==14 or pp==28 or pp==18 or pp==24:
        continue
        
    if pp==6 or pp==32 or pp==12:
        continue
        
    counter+=1
    
    crystalname=crystal(pp)

    columnname=crystalname+'_'+suffix+'_RDF'
    
    crystalnames_df.append(crystalname)

    print(columnname)
    
    RDF = df_intch[columnname]
    
    # print(len(RDF))

    BADER= df_bader.loc[counter-1]
    LATTICE= df_lattice.loc[counter-1]
    STRUCTURAL=df_structural.loc[pp-1]
    VALENCE=df_valence.loc[pp-1]
    
    RDF = np.asarray(RDF)
    # RDF = RDF[np.logical_not(np.isnan(RDF))]
    RDF =np.roll(RDF, 50)
    
    MAX_rdf = np.amax(RDF)
    # MIN_rdf = np.amin(RDF)
            
    # ------ signal info

    peaks_RDF, peaks_RDF_properties = find_peaks(RDF, height=0)
    # print("\n peaks_RDF, peaks_RDF_properties",peaks_RDF, peaks_RDF_properties)
    peaks_RDF_widths=scipy.signal.peak_widths(RDF, peaks_RDF, rel_height=0.5, prominence_data=None, wlen=None)
    # print("\n peaks_RDF_widths",peaks_RDF_widths)

    all_heights=np.asarray(peaks_RDF_properties['peak_heights'])
    max_Height=max(all_heights)
    index=np.where(all_heights==max_Height)

    peak_width=peaks_RDF_widths[0][index]
    # print("\n peak_width",peak_width)
    # peak1_width_half_height=peaks_RDF_widths[1][0]
    # print("\n peak1_width_half_height",peak1_width_half_height)

    # peak2_pos=Rr_axis[peaks_RDF][1]
    # peak2_height=peaks_RDF_properties['peak_heights'][1]
    # peak2_width=peaks_RDF_widths[0][1]
    # peak2_width_half_height=peaks_RDF_widths[1][1]

    
    row=[element(crystalname).name,
        element(crystalname).atomic_number,
        element(crystalname).atomic_radius,
        element(crystalname).atomic_volume,
        element(crystalname).atomic_weight,
        element(crystalname).dipole_polarizability,
        element(crystalname).period,
        element(crystalname).boiling_point,
        element(crystalname).c6,
        element(crystalname).covalent_radius_bragg,
        element(crystalname).covalent_radius_cordero,
        element(crystalname).covalent_radius_pyykko,
        element(crystalname).covalent_radius_pyykko_double,
        element(crystalname).covalent_radius_pyykko_triple,
        element(crystalname).c6_gb,
        element(crystalname).density,
        element(crystalname).dipole_polarizability_unc,
        element(crystalname).electron_affinity,
        element(crystalname).evaporation_heat,
        element(crystalname).fusion_heat,
        element(crystalname).gas_basicity,
        element(crystalname).glawe_number,
        element(crystalname).heat_of_formation,
        element(crystalname).ionic_radii[0].ionic_radius, # there is also [1]
        element(crystalname).melting_point,
        element(crystalname).mendeleev_number,
        element(crystalname).metallic_radius,
        element(crystalname).metallic_radius_c12,
        element(crystalname).proton_affinity,
        element(crystalname).specific_heat,
        element(crystalname).thermal_conductivity,
        element(crystalname).vdw_radius,
        element(crystalname).vdw_radius_alvarez,
        element(crystalname).vdw_radius_batsanov,
        element(crystalname).vdw_radius_bondi,
        element(crystalname).vdw_radius_dreiding,
        element(crystalname).vdw_radius_mm3,
        element(crystalname).vdw_radius_rt,
        element(crystalname).vdw_radius_truhlar,
        element(crystalname).vdw_radius_uff,
        VALENCE['VALENCE'],
        BADER['Ch(M)'], #df_bader_OCTA['Ch(M)'],
        BADER['Ch(H)'],#df_bader_OCTA['Ch(H)'], #should address to the right element when calling 
        BADER['Rad_M'],#df_bader_OCTA['Rad_M'],
        BADER['Rad_H'],#df_bader_OCTA['Rad_H'],
        BADER['V_M'],#df_bader_OCTA['V_M'],
        BADER['V_H'],
        LATTICE['Lat'],
        STRUCTURAL['B.M.'],
        STRUCTURAL['Y.M.'],
        STRUCTURAL['S.M.'],
        STRUCTURAL['P.R.'],      
        peak_width[0],
        MAX_rdf,     
        ]
    
    data.append(row)
    
    df = pd.DataFrame(data, columns = ['Name', 'Atomic Number', 'Atomic Radius', 'Atomic Volume','atomic_weight',
        'dipole_polarizability',
        'period',
        'boiling_point',
        'c6',
        'covalent_radius_bragg',
        'covalent_radius_cordero',
        'covalent_radius_pyykko',
        'covalent_radius_pyykko_double',
        'covalent_radius_pyykko_triple',
        'c6_gb',
        'density',
        'dipole_polarizability_unc',
        'electron_affinity',
        'evaporation_heat',
        'fusion_heat',
        'gas_basicity',
        'glawe_number',
        'heat_of_formation',
        'ionic_radii',
        'melting_point',
        'mendeleev_number',
        'metallic_radius',
        'metallic_radius_c12',                       
        'proton_affinity',
        'specific_heat',                          
        'thermal_conductivity',
        'vdw_radius',
        'vdw_radius_alvarez',
        'vdw_radius_batsanov',
        'vdw_radius_bondi',
        'vdw_radius_dreiding',
        'vdw_radius_mm3',
        'vdw_radius_rt',
        'vdw_radius_truhlar',
        'vdw_radius_uff',
        'VALENCE',
        'Ch(M)',
        'Ch(H)',
        'Rad_M',
        'Rad_H',
        'V_M',
        'V_H',
        'Lat',
        'BULKM',
        'YOUNGM', 
        'SHEARM',
        'POISSONR',
        'peak_width',
        'MAX_rdf',
      ])

    # for ii in range (0,174):
    #     print(RDF[ii])
    #     df.insert(len(df.columns),'**RDF_'+str(ii)+'**',RDF[ii])

df

    


# R_cont=[]
# for nx in range (0,101):
#     for ny in range(0,101):
#         for nz in range(0,101):
            
#             R_cont.append(np.sqrt(nx**2+ny**2+nz**2))

# R_cont=np.array(R_cont)
# R=R_cont.astype(int)
# Rr=np.unique(R_cont.ravel()) 
# Rrmax=max(Rr)
# nr = np.bincount(R)
# Rr_axis=np.linspace(0,Rrmax,len(nr))

# #!pip install pymatgen
# #import pymatgen.core as mg
# #df['ionic_radii']
# print(df[['Name','peak1_pos','peak1_height','peak2_pos','peak2_height']])

df.shape

df.describe()

df.head()

df.to_csv('OCTA_DATASET.csv',index=False)

X=df.drop(columns = 'Name',axis=1)
X.isna().sum()
